\name{woRdsmith-package}
\alias{woRdsmith-package}
\alias{woRdsmith}
\docType{package}
\title{
Quick little typing game written in R
}
\description{
Typing game that works in the terminal version of R. Windows GUI and Rstudio do not work
}
\details{
\tabular{ll}{
Package: \tab woRdsmith\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2018-06-23\cr
License: \tab None\cr
}
}